# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-Lago API (Numaris-native)",
    "version": "2.0.0",
    "description": "Motor de facturación domain-specific para Numaris. Entidades: Customer, Tax, Service (colapsa Plan+Sub+Charge+BM+AddOn), Unit (materializada), EventLog (audit), Invoice, CreditNote.\n\n**Auth**: Bearer en todos los `/api/v1/*` excepto health/openapi/discovery. Callbacks NetSuite usan HMAC-SHA256 (`X-NetSuite-Signature`).\n\n**Idempotency** en POST /invoices y POST /credit_notes vía header `Idempotency-Key` + `metadata.idempotency_key`."
  },
  "servers": [
    { "url": "https://billing2-production.up.railway.app", "description": "Railway production" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "NetSuiteSignature": { "type": "apiKey", "in": "header", "name": "X-NetSuite-Signature" }
    },
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "external_id": { "type": "string" },
          "sequential_id": { "type": "integer" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": ["string", "null"] },
          "phone": { "type": ["string", "null"] },
          "tax_identification_number": { "type": ["string", "null"] },
          "address_line1": { "type": ["string", "null"] },
          "city": { "type": ["string", "null"] },
          "country": { "type": ["string", "null"] },
          "currency": { "type": "string" },
          "timezone": { "type": ["string", "null"] },
          "applicable_timezone": { "type": "string" },
          "metadata": { "type": "object" },
          "taxes": { "type": "array", "items": { "$ref": "#/components/schemas/Tax" } }
        }
      },
      "CustomerWriteable": {
        "type": "object",
        "required": ["external_id", "name", "currency"],
        "properties": {
          "external_id": { "type": "string" },
          "name": { "type": "string" },
          "currency": { "type": "string", "example": "MXN" },
          "country": { "type": "string", "example": "MX" },
          "timezone": { "type": "string", "example": "America/Mexico_City" },
          "tax_identification_number": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address_line1": { "type": "string" },
          "city": { "type": "string" },
          "metadata": { "type": "object" },
          "tax_codes": { "type": "array", "items": { "type": "string" }, "description": "Reemplazo total (D2)." }
        }
      },
      "Tax": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "rate": { "type": "number", "example": 16.0 }
        }
      },
...
      "CreditNoteWriteable": {
        "type": "object",
        "required": ["invoice_id", "reason", "items"],
        "properties": {
          "invoice_id": { "type": "string", "format": "uuid" },
          "reason": { "type": "string", "enum": ["duplicated_charge", "product_unsatisfactory", "order_change", "order_cancellation", "fraudulent_charge", "other"] },
          "description": { "type": "string" },
          "credit_amount_cents": { "type": "integer" },
          "refund_amount_cents": { "type": "integer" },
          "items": { "type": "array", "items": { "type": "object", "required": ["fee_id", "amount_cents"], "properties": { "fee_id": { "type": "string", "format": "uuid" }, "amount_cents": { "type": "integer", "minimum": 1 } } } }
        }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "summary": "Create / upsert customer por external_id (D1)",
        "tags": ["customers"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "customer": { "$ref": "#/components/schemas/CustomerWriteable" } } } } } },
        "responses": { "200": { "description": "Customer", "content": { "application/json": { "schema": { "type": "object", "properties": { "customer": { "$ref": "#/components/schemas/Customer" } } } } } } }
      },
      "get": {
        "summary": "List customers (paginado)",
        "tags": ["customers"],
        "parameters": [
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": { "200": { "description": "Lista" } }
      }
    },
    "/api/v1/customers/{externalId}": {
      "get": { "summary": "Read customer", "tags": ["customers"], "parameters": [{ "name": "externalId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Customer" }, "404": { "description": "customer_not_found" } } },
      "delete": { "summary": "Delete customer (409 si tiene services activos)", "tags": ["customers"], "parameters": [{ "name": "externalId", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" }, "409": { "description": "customer_has_active_services" } } }
    },
...
    "/api/v1/credit_notes": {
      "post": {
        "summary": "Create credit note (solo sobre invoice finalized+confirmed)",
        "tags": ["credit_notes"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "credit_note": { "$ref": "#/components/schemas/CreditNoteWriteable" } } } } } },
        "responses": { "200": { "description": "CN" }, "422": { "description": "invoice_not_confirmed" } }
      },
      "get": { "summary": "List credit notes", "tags": ["credit_notes"], "parameters": [{ "name": "customer_external_id", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Lista" } } }
    },
    "/api/v1/credit_notes/{id}": {
      "get": { "summary": "Read credit note", "tags": ["credit_notes"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "CN" } } }
    },
...
  }
}